fixed check for missing textures
img.has_data is True only if the data of the image has been loaded into memory, which might or might not be the case for available textures.
So instead we ask for img.size, which forces loading the image.
(Afterwards we could use img.has_data, but img.size == (0, 0) is also a clear sign for missing data so we use that)