Adding SpotLight to Lighting kit (#219)
* Adding Spotlight requirements

Added the spot_blend and spot_size parameters into the lights.py file.

Spot Blend is the roughness of the beam
Spot Size is the radius in deg 0 - 180

Blender.py was changed to add the needed parameters inside the blender handler. I think I added the correct identifiers and what not. I just copied the standard implementation as it was written.

The SpotLight was added in with the other forms of Lights

----------
Still need to look at the colors of the lights. Maybe refer to the issue in the main github of Kubric regarding the Color Class and its implementation.

* Update as per notes

Awesome notes from Qwlouse! can be found at PR #219

* As per #225

As per #225 fix implemented

* Update kubric/renderer/blender.py

Co-authored-by: Klaus Greff <klausg@google.com>

* Import Numpy

Bug Fix noted by Qwlouse

Co-authored-by: Klaus Greff <klausg@google.com>