Fix bug when loading glbs with multiple meshes. (#310)
The previous code was deleting all objects that were not of type "MESH" and then join all 'MESH' types. Doing it that way seemed to cause objects to 'split up' as they seem to have different transforms and their different parts are distributed all over the scene.

The new code selects all objects of type "MESH" without removal of any objects. This seems to fix the issue.

PiperOrigin-RevId: 611289628

Co-authored-by: kubric-team <kubric@google.com>